unify block create and update for rewards
